requirements.txt
tensorflow
flask
pillow
numpy
matplotlib
scikit-learn